% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{api_data_call}
\alias{api_data_call}
\title{api_data_call}
\usage{
api_data_call(unit = "", .project_path = "")
}
\arguments{
\item{.project_name}{}
}
\description{
[AUX.] Wrapper that calls to the current state of the coding task to feed to
the API upon a GET from the UI client.
}
