% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{check_db_exists}
\alias{check_db_exists}
\title{check_db_exists}
\usage{
check_db_exists(.project_path = "")
}
\arguments{
\item{path}{}
}
\value{

}
\description{
[Aux.] Check if a data base file exist
}
\examples{

check_db_exists()

}
