% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{generate_module}
\alias{generate_module}
\title{generate_module()}
\usage{
generate_module(.data, variable_name = "", caption = "", ...)
}
\arguments{
\item{variable_name}{}

\item{...}{}
}
\description{
Build field inputs for coding entry. The function will run all
}
