% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{import_data_state}
\alias{import_data_state}
\title{import_data_state}
\usage{
import_data_state(.project_path = "", empty_value_placeholder = "")
}
\arguments{
\item{.project_name}{initial name of the data_base build.}
}
\value{
exports all units along with the current state of the projec
}
\description{
[AUX.] Function reads in and reports on the state of the established coding
task
}
