% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{map_input}
\alias{map_input}
\title{map_input}
\usage{
map_input(con, ...)
}
\arguments{
\item{con}{sqlite connection}

\item{...}{all entry fields passed from generate module.}
}
\value{

}
\description{
Auxiliary function that generates instructions for the UI application.
}
