% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{populate_source}
\alias{populate_source}
\title{populate_source}
\usage{
populate_source(source_name = "", expr)
}
\arguments{
\item{source_name}{}

\item{expr}{}
}
\value{
js code snippet with the appropriate source name
}
\description{
[Aux] populate internal source code for js argument.
}
