% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{pull_data}
\alias{pull_data}
\title{pull_data}
\usage{
pull_data(.project_path = "")
}
\arguments{
\item{.project_path}{}
}
\description{
Posterior function that call from the data base to make a distinct data frame
that contains all fields from all coded variables from the application. The
function allows users to draw the most recent state of the coding task to
convert into a usable data frame for analysis
}
\examples{

# Must point to the project path created by qualify()
pull_data(.project_path = "~/Desktop/test_project/")

}
