% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{pull_timeline}
\alias{pull_timeline}
\title{pull_timeline}
\usage{
pull_timeline(.project_path, round_date = "minute")
}
\arguments{
\item{.project_path}{}
}
\description{
Pull in timeline information about the coding timeline.
}
