% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{qualify}
\alias{qualify}
\title{qualify}
\usage{
qualify(project_name = "test_db",
  project_path = "~/Desktop/test_project", unit_of_analysis = NULL)
}
\arguments{
\item{project_name}{}

\item{project_path}{}

\item{unit_of_analysis}{}
}
\description{
Master function that initializes the data object
}
