% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{sql_instance}
\alias{sql_instance}
\title{sql_instance}
\usage{
sql_instance(path = "")
}
\arguments{
\item{db_name}{Name of the data base structure}
}
\value{
Connection to SQLite DB
}
\description{
[AUX] Generate/call existing SQL database.
}
