% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualify.R
\name{upload_data}
\alias{upload_data}
\title{upload_data}
\usage{
upload_data(entry, .project_path = "")
}
\arguments{
\item{.project_name}{}
}
\description{
[AUX.] Wrapper that saves user input to the existing database structure after
receiving a post request from the client.
}
